### 2023-09-20 #5857 @editorialbot
[PRE REVIEW]: Pyrokinetics - A Python library to standardise gyrokinetic analysis
**Submitting author:** @bpatel2107 (<a href="http://orcid.org/0000-0003-0121-1187">Bhavin Patel</a>)
**Repository:** https://github.com/pyro-kinetics/pyrokinetics
**Branch with paper.md** (empty if default branch): docs/joss_paper
**Version:** v0.5.0
