+++ pyrokinetics/gk_code/gene.py
@@ -346,6 +346,7 @@ class GKInputGENE(GKInput):
         )
         domega_drho = self.data["geometry"]["q0"] / rho * external_contr["ExBrate"]
 
+        ion_names = []
         # Load each species into a dictionary
         for i_sp in range(self.data["box"]["n_spec"]):
             species_data = CleverDict()
@@ -375,7 +376,8 @@ class GKInputGENE(GKInput):
                 ) * (ureg.vref_nrl / self.lref_gene)
             else:
                 ion_count += 1
-                name = f"ion{ion_count}"
+                name = gene_data["name"]
+                ion_names.append(name)
                 species_data.nu = None
 
             species_data.name = name
@@ -395,7 +397,7 @@ class GKInputGENE(GKInput):
         me = local_species.electron.mass
 
         for ion in range(ion_count):
-            key = f"ion{ion + 1}"
+            key = ion_names[ion]
 
             nion = local_species[key]["dens"]
             tion = local_species[key]["temp"]
@@ -718,7 +720,7 @@ class GKOutputReaderGENE(AbstractFileReader):
         # Assign units and return GKOutput
         field_dims = ("theta", "kx", "ky", "time")
         flux_dims = ("field", "species", "time")
-        moment_dims = ("field", "species", "time")
+        moment_dims = ("theta", "kx", "species", "ky", "time")
         return GKOutput(
             coords=Coords(
                 time=coords["time"],
@@ -808,6 +810,49 @@ class GKOutputReaderGENE(AbstractFileReader):
                 files.update({"field": dirname / f"field{delimiter}{suffix}.h5"})
         return files
 
+    @staticmethod
+    def _get_gene_mom_files(filename: PathLike, files: Dict, species_names) -> Dict[str, Path]:
+        """
+        Given a directory name, looks for the files filename/parameters_0000,
+        filename/field_0000 and filename/nrg_0000.
+        If instead given any of the files parameters_####, field_#### or nrg_####,
+        looks up the rest of the files in the same directory.
+        """
+        filename = Path(filename)
+        prefixes = [f"mom_{species_name}" for species_name in species_names]
+        if filename.is_dir():
+            # If given a dir name, looks for dir/parameters_0000
+            dirname = filename
+            dat_matches = np.any(
+                [Path(filename / f"{p}.dat").is_file() for p in prefixes]
+            )
+            if dat_matches:
+                suffix = "dat"
+                delimiter = "."
+            else:
+                suffix = "0000"
+                delimiter = "_"
+        else:
+            # If given a file, searches for all similar GENE files in that file's dir
+            dirname = filename.parent
+            # Ensure provided file is a GENE file (fr"..." means raw format str)
+            matches = [re.search(rf"^{p}_\d{{4}}$", filename.name) for p in prefixes]
+            if not np.any(matches):
+                raise RuntimeError(
+                    f"GKOutputReaderGENE: The provided file {filename} is not a GENE "
+                    "output file."
+                )
+            suffix = filename.name.split("_")[1]
+            delimiter = "_"
+
+        # Get all files in the same dir
+        for prefix in prefixes:
+            if (dirname / f"{prefix}{delimiter}{suffix}").exists():
+                files[prefix] = dirname / f"{prefix}{delimiter}{suffix}"
+
+        return files
+
+
     def verify_file_type(self, filename: PathLike):
         self._get_gene_files(filename)
 
@@ -839,6 +884,9 @@ class GKOutputReaderGENE(AbstractFileReader):
             input_str = f.read()
         gk_input = GKInputGENE()
         gk_input.read_str(input_str)
+
+        species_names = gk_input.get_local_species().names
+        files = cls._get_gene_mom_files(filename,files, species_names)
[…]
