+++ docs/joss/paper.bib
@@ -1,29 +1,37 @@
 @article{giacomin:2023a,
-  title={Numerical investigation of microtearing modes in the core of experimentally relevant spherical tokamak scenarios},
-  author={Giacomin, M and Dickinson, D and Kennedy, D and Patel, BS and Roach, CM},
-  journal={arXiv preprint arXiv:2303.02379},
-  year={2023}
+  title={Nonlinear microtearing modes in MAST and their stochastic layer formation},
+  author={Giacomin, Maurizio and Dickinson, David and Kennedy, Daniel and Patel, BS and Roach, CM},
+  journal={Plasma Physics and Controlled Fusion},
+  volume={65},
+  number={9},
+  pages={095019},
+  year={2023},
+  publisher={IOP Publishing},
+  doi={10.1088/1361-6587/aceb89},
 }
 
 @article{giacomin:2023b,
   title={Electromagnetic gyrokinetic instabilities in the Spherical Tokamak for Energy Production (STEP) part II: transport and turbulence},
   author={Giacomin, Maurizio and Kennedy, Daniel and Casson, Francis J and Dickinson, David and Patel, Bhavin S and Roach, Colin M and others},
   journal={arXiv preprint arXiv:2307.01669},
-  year={2023}
+  year={2023},
+  doi={10.48550/arXiv.2307.01669},
 }
 
 @article{kennedy:2023,
   title={Electromagnetic gyrokinetic instabilities in the Spherical Tokamak for Energy Production (STEP) part I: linear physics and sensitivity},
   author={Kennedy, Daniel and Giacomin, Maurizio and Casson, Francis J and Dickinson, David and Hornsby, William A and Patel, Bhavin S and Roach, Colin M},
   journal={arXiv preprint arXiv:2307.01670},
-  year={2023}
+  year={2023},
+  doi={10.48550/arXiv.2307.01670},
 }
 
 @article{hornsby:2023,
   title={Gaussian Process Regression models for the properties of micro-tearing modes for spherical tokamaks},
   author={W.A Hornsby, A. Gray, J. Buchanan, B.S. Patel, D. Kennedy, F.J. Casson, C. Roach, M. B. Lykkegaard, H.Nguyen, N. Papadimas, B. Fourcin, J. Hart},
-  journal={Phys. of Plasma - to be submitted},
-  year={2023}
+  journal={arXiv preprint arXiv:2309.09785},
+  year={2023},
+  doi={10.48550/arXiv.2309.09785},
 }
 
 @article{imbeaux:2015,
@@ -34,7 +42,8 @@
   number={12},
   pages={123006},
   year={2015},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi={10.1088/0029-5515/55/12/123006},
 }
 
 @article{pankin:2004,
@@ -45,7 +54,8 @@
   number={3},
   pages={157--184},
   year={2004},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.cpc.2003.11.002},
 }
 
 @article{cenacchi:1988,
