### 2023-09-22 52f69ca
Set moments from correct values in mom files
