### 2023-09-22 2cfd850
Revert kx grid
