### 2023-10-04 3cd2187
Add system to register plugins via entry points
