+++ pyrokinetics/kinetics/__init__.py
@@ -1,8 +1,14 @@
-# Import KineticsReaders
 from .kinetics import Kinetics, read_kinetics, supported_kinetics_types
+
+# Import each built-in reader to register them with Kinetics
 from .scene import KineticsReaderSCENE  # noqa
 from .jetto import KineticsReaderJETTO  # noqa
 from .transp import KineticsReaderTRANSP  # noqa
 from .pfile import KineticsReaderpFile  # noqa
+
+# Register external plugins with Kinetics
+from ..plugins import register_file_reader_plugins
+
+register_file_reader_plugins("Kinetics", Kinetics)
 
 __all__ = ["Kinetics", "read_kinetics", "supported_kinetics_types"]
