+++ pyrokinetics/equilibrium/__init__.py
@@ -6,9 +6,14 @@ from .equilibrium import (
 )
 from .flux_surface import FluxSurface
 
-# Import each reader to register them with the factory
+# Import each built-in reader to register them with Equilibrium
 from . import geqdsk  # noqa
 from . import transp  # noqa
+
+# Register external plugins with Equilibrium
+from ..plugins import register_file_reader_plugins
+
+register_file_reader_plugins("Equilibrium", Equilibrium)
 
 __all__ = [
     "Equilibrium",
