### 2023-10-04 7f64d07
Add system to register plugins via entry points
