### 2023-10-04 4b40220
Update file readers/readables + factory methods
Use __init_subclass__ in place of awkward decorator + superclass
method. Simplifies factory interface.
