+++ pyrokinetics/plugins.py
@@ -3,23 +3,23 @@ This modules contains utility functions for adding plugins to Pyrokinetics.
 This is achieved using entry-points.
 
 If you want to register your own :class:`~pyrokinetics.equilibrium.Equilibrium`
-reader, it should inherit :class:`pyrokinetics.file_utils.AbstractFileReader`,
+reader, it should inherit :class:`pyrokinetics.file_utils.FileReader`,
 and its ``read_from_file()`` function should return an ``Equilibrium``. To add
 this plugin to Pyrokinetics, you should add the following to your
 ``pyproject.toml`` file::
 
-    [project.entry-points."pyrokinetics.Equilibrium"]
+    [project.entry-points."pyrokinetics.equilibrium"]
     my_eq = "my_project.my_module:MyEqReader"
 
 This will register the class ``MyEqReader``, and within Pyrokinetics the
-equilibrium type will be ``"my_eq"``. Note that here, ``"pyrokinetics.Equilibrium"`` is 
-an entry point group name, not a class. The group names for each Pyrokinetics
+equilibrium type will be ``"my_eq"``. Note that here, ``"pyrokinetics.equilibrium"`` is 
+an entry point group name, not a module. The group names for each Pyrokinetics
 file reader are:
 
-- ``"pyrokinetics.GKInput"``
-- ``"pyrokinetics.GKOutput"``
-- ``"pyrokinetics.Equilibrium"``
-- ``"pyrokinetics.Kinetics"``
+- ``"pyrokinetics.gkinput"``
+- ``"pyrokinetics.gkoutput"``
+- ``"pyrokinetics.equilibrium"``
+- ``"pyrokinetics.kinetics"``
 
 For more information, please see:
 
@@ -30,37 +30,20 @@ For more information, please see:
 """
 
 from importlib.metadata import entry_points
+from textwrap import dedent
 from typing import Type
 
-from .file_utils import AbstractFileReader, ReadableFromFileMixin
+from .file_utils import FileReader, ReadableFromFile
 
-__all__ = ["register_file_reader_plugins", "PluginError"]
-
-
-class PluginError(TypeError):
-    """
-    Variant on ``TypeError`` to be thrown when plugins cannot be loaded.
-
-    Parameters
-    ----------
-    plugin_cls
-        The class the user is trying to register as a plugin.
-    super_cls
-        The Pyrokinetics super class that the plugin class inherits.
-    """
-
-    def __init__(self, plugin_cls: Type, super_cls: Type) -> None:
-        super().__init__(
-            f"{plugin_cls.__qualname__} must subclass {super_cls.__qualname__}"
-        )
+__all__ = ["register_file_reader_plugins"]
 
 
 def register_file_reader_plugins(
-    group_name: str, Readable: Type[ReadableFromFileMixin]
+    group_name: str, Readable: Type[ReadableFromFile]
 ) -> None:
     """
     Defines an entry point group for classes that implement the
-    :class:`~pyrokinetics.file_utils.AbstractFileReader` and registers any
+    :class:`~pyrokinetics.file_utils.FileReader` and registers any
     user-defined plugins with that group.
 
     Parameters
@@ -84,7 +67,21 @@ def register_file_reader_plugins(
 
     # Register all plugins in the user's environment
     for entry in group:
+        # Simply loading the entry point should register the plugin if it
+        # inherits FileReader
         cls = entry.load()
-        if not issubclass(cls, AbstractFileReader):
-            raise PluginError(cls, AbstractFileReader)
-        Readable.reader(entry.name)(cls)
+        # Check that it is of the correct type
+        if not issubclass(cls, FileReader):
+            raise TypeError(
+                f"Plugin class {cls.__qualname__} should subclass {cls.__qualname__}"
+            )
+        # Check that the registered file type matches the entry point name
[…]
