+++ docs/tutorials/file_readers.rst
@@ -43,16 +43,11 @@ To begin, we must mark ``Foo`` as being readable. This is achieved as so:
     import numpy as np
     from numpy.typing import ArrayLike
 
-    # These components are needed to make something 'readable'
-    from pyrokinetics.file_utils import (
-        readable_from_file,
-        ReadableFromFileMixin,
-    )
+    # The base class ReadableFromFile is needed to mark
+    # a class as a 'readable'
+    from pyrokinetics.file_utils import ReadableFromFile
 
-    # readable_from_file is used as a decorator, while
-    # ReadableFromFileMixin is used as a super-class.
-    @readable_from_file
-    class Foo(ReadableFromFileMixin):
+    class Foo(ReadableFromFile):
 
         # __init__ should take raw data -- not a file path
         def __init__(self, x: ArrayLike, y: ArrayLike):
@@ -64,48 +59,36 @@ To begin, we must mark ``Foo`` as being readable. This is achieved as so:
             self._y = np.asarray(y)
 
 In order to make a readable class, we have sub-classed
-:class:`~pyrokinetics.file_utils.ReadableFromFileMixin` and decorated the class with
-:func:`~pyrokinetics.file_utils.readable_from_file`. Both of these components are
-necessary:
+:class:`~pyrokinetics.file_utils.ReadableFromFile`. This adds the following
+classmethods:
 
-- :class:`~pyrokinetics.file_utils.ReadableFromFileMixin` contains the classmethods:
+- ``from_file``: Used as an alternative constructor. Creates a readable from a file.
 
-   - ``from_file``: Used as an alternative constructor. Creates a readable from a file.
-
-   - ``supported_file_types``: Returns a list of file types that ``from_file`` can read.
-
-   - ``reader``: A decorator used to tag other classes as 'readers' and associate them
-     with this readable.
-
-- :func:`~pyrokinetics.file_utils.readable_from_file` adds class-level attributes that
-  are used by the aforementioned classmethods.
+- ``supported_file_types``: Returns a list of file types that ``from_file`` can read.
 
 Having defined a 'readable' class, we can now define an associated reader:
 
 .. code-block:: python
 
     # file: foo_csv_reader.py
-    from pyrokinetics.file_utils import AbstractFileReader
+    from pyrokinetics.file_utils import FileReader
     from .foo import Foo
 
-    @Foo.reader("csv")
-    class FooReaderCSV(AbstractFileReader):
+    class FooReaderCSV(FileReader, file_type="csv", reads=Foo):
         ...
 
-Again, we have sub-classed a class from :mod:`~pyrokinetics.file_utils` and wrapped the
-class with a decorator:
+Again, we have sub-classed a class from :mod:`~pyrokinetics.file_utils`. We have also
+added two required keyword arguments to the class:
 
-- The decorator ``Foo.reader`` was added to ``Foo`` by
-  :class:`~pyrokinetics.file_utils.ReadableFromFileMixin`. This 'registers' the reader
-  with an associated readable via a key. This key should be the name of the file type we
-  wish to read, or the name of the software that generated the file.
+- :class:`~pyrokinetics.file_utils.FileReader` defines the abstract method
+  :meth:`~pyrokinetics.file_utils.FileReader.read_from_file` and the method
+  :meth:`~pyrokinetics.file_utils.FileReader.verify_file_type`. This means that
+  sub-classes must provide a definition of ``read_from_file()``, or else Python will 
+  throw an error. The former method is used to read/process data from files, while the
+  latter is used to determine whether a file is of the correct type.
 
-- :class:`~pyrokinetics.file_utils.AbstractFileReader` defines abstract methods
[…]
