### 2023-10-05 f331417
Add equilibrium plugin test
