### 2023-10-05 f3ffcb3
Add equilibrium plugin test
