+++ pyproject.toml
@@ -57,6 +57,7 @@ docs = [
 tests = [
     "pytest >= 3.3.0",
     "pytest-cov",
+    "pyrokinetics-plugin-examples == 0.1.0",
 ]
 
 [project.urls]
