### 2023-10-05 3bb782d
Update test workflow to get all test dependencies
