+++ .github/workflows/tests.yml
@@ -25,8 +25,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install .
+        pip install .[tests]
     - name: Test with pytest
       run: |
-        pip install pytest
-        pytest -v
+        pytest --cov --cov-report=xml --cov-report=term -vv
