+++ .github/workflows/ci.yml
@@ -15,11 +15,9 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install .
+        pip install .[tests]
     - name: Run tests and collect coverage
       run: |
-        pip install pytest
-        pip install pytest-cov
         pytest --cov --cov-report=xml
     - name: Upload coverage reports to Codecov with GitHub Action
       uses: codecov/codecov-action@v3
