+++ pyproject.toml
@@ -82,3 +82,8 @@ git_describe_command = "git describe --dirty --tags --long --match v* --first-pa
 
 [tool.setuptools.package-data]
 pyrokinetics = ["templates/*"]
+
+[tool.pytest.ini_options]
+addopts = [
+    "--import-mode=importlib",
+]
