+++ .github/workflows/tests.yml
@@ -9,7 +9,9 @@ on:
       - '**.py'
 
 jobs:
-  pytest:
+
+  run-tests:
+    name: Run tests
     runs-on: ubuntu-latest
     strategy:
       fail-fast: false
@@ -17,16 +19,35 @@ jobs:
         python-version: ["3.8", "3.9", "3.10"]
 
     steps:
-    - uses: actions/checkout@v2
+    - uses: actions/checkout@v4
     - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v1
+      uses: actions/setup-python@v4
       with:
         python-version: ${{ matrix.python-version }}
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install .
+        python -m pip install .[tests]
     - name: Test with pytest
       run: |
-        pip install pytest
-        pytest -v
+        pytest --cov=pyrokinetics --cov-report=xml --cov-report=term tests
+    - name: Upload coverage artifacts
+      uses: actions/upload-artifact@v3
+      with:
+        name: ${{ format('coverage-python-{0}, matrix.python-version) }}
+        path: coverage.xml
+
+  code-coverage:
+    name: Code coverage
+    needs:
+    - run-tests
+    runs-on: ubuntu-latest
+
+    steps:
+    - name: Download coverage report
+      uses: actions/download-artifact@v3
+      with:
+        name: "coverage-python-3.9"
+        path: coverage.xml
+    - name: Upload coverage reports to Codecov
+      uses: codecov/codecov-action@v3
