+++ tests/gk_code/test_gk_input_gene.py
@@ -9,6 +9,7 @@ import numpy as np
 import pytest
 
 import sys
+
 docs_dir = Path(__file__).parent.parent.parent / "docs"
 sys.path.append(str(docs_dir))
 from examples import example_JETTO  # noqa
