### 2023-10-05 6e9cdf2
Undo rename of pytest workflow job
