+++ .github/workflows/tests.yml
@@ -10,8 +10,7 @@ on:
 
 jobs:
 
-  run-tests:
-    name: Run tests
+  pytest:
     runs-on: ubuntu-latest
     strategy:
       fail-fast: false
