+++ .github/workflows/flake8.yml
@@ -26,7 +26,7 @@ jobs:
       # is going to solve anyway
       - name: Run black
         run: |
-          black pyrokinetics
+          black src tests
       - name: Run flake8
         uses: suo/flake8-github-action@releases/v1
         with:
