### 2023-10-05 604bd8b
Move .coveragerc to pyproject.toml, fix relative paths
