### 2023-10-05 05fbb89
Edit omitted coverage paths
