+++ .github/workflows/tests.yml
@@ -43,6 +43,7 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
+    - uses: actions/checkout@v4
     - name: Download coverage report
       uses: actions/download-artifact@v3
       with:
