### 2023-10-06 f5120b0
Add Python 3.11 tests, use 3.10 coverage report with Codecov
