+++ .github/workflows/tests.yml
@@ -15,7 +15,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.8", "3.9", "3.10"]
+        python-version: ["3.8", "3.9", "3.10", "3.11"]
 
     steps:
     - uses: actions/checkout@v4
@@ -47,7 +47,7 @@ jobs:
     - name: Download coverage report
       uses: actions/download-artifact@v3
       with:
-        name: "coverage-python-3.9"
+        name: "coverage-python-3.10"
         path: coverage.xml
     - name: Upload coverage reports to Codecov
       uses: codecov/codecov-action@v3
