### 2023-10-06 1e1153d
Add additional plugin tests
