+++ pyproject.toml
@@ -57,7 +57,7 @@ docs = [
 tests = [
     "pytest >= 3.3.0",
     "pytest-cov",
-    "pyrokinetics-plugin-examples == 0.1.0",
+    "pyrokinetics-plugin-examples == 0.2.0",
 ]
 
 [project.urls]
