### 2023-10-06 ca8de9f
Add additional plugin tests
