+++ tests/test_plugins.py
@@ -3,7 +3,21 @@ from pathlib import Path
 import pytest
 from pyrokinetics_plugin_examples import make_plugin_example_file
 
-from pyrokinetics import Equilibrium, supported_equilibrium_types, read_equilibrium
+from pyrokinetics.equilibrium import (
+    Equilibrium,
+    supported_equilibrium_types,
+    read_equilibrium,
+)
+from pyrokinetics.kinetics import Kinetics, supported_kinetics_types, read_kinetics
+from pyrokinetics.gk_code import (
+    GKInput,
+    GKOutput,
+    supported_gk_input_types,
+    supported_gk_output_types,
+    read_gk_input,
+    read_gk_output,
+)
+from pyrokinetics.normalisation import SimulationNormalisation
 
 
 @pytest.fixture
@@ -24,3 +38,19 @@ def plugin_file(tmp_path: Path) -> Path:
 def test_equilibrium_plugin(plugin_file: Path) -> None:
     assert "_test" in supported_equilibrium_types()
     assert isinstance(read_equilibrium(plugin_file), Equilibrium)
+
+
+def test_kinetics_plugin(plugin_file: Path) -> None:
+    assert "_test" in supported_kinetics_types()
+    assert isinstance(read_kinetics(plugin_file), Kinetics)
+
+
+def test_gk_input_plugin(plugin_file: Path) -> None:
+    assert "_test" in supported_gk_input_types()
+    assert isinstance(read_gk_input(plugin_file), GKInput)
+
+
+def test_gk_output_plugin(plugin_file: Path) -> None:
+    assert "_test" in supported_gk_output_types()
+    norm = SimulationNormalisation("_test", convention="pyrokinetics")
+    assert isinstance(read_gk_output(plugin_file, norm), GKOutput)
