### 2023-10-06 a709584
Move to src layout, anticipating merge with PR 258
