+++ src/pyrokinetics/gk_code/gx.py
@@ -1,0 +1,811 @@
+import warnings
+from copy import copy
+from itertools import product
+from pathlib import Path
+from typing import Any, Dict, Optional, Tuple
+
+import numpy as np
+import xarray as xr
+from cleverdict import CleverDict
+
+from ..constants import pi, sqrt2
+from ..local_geometry import LocalGeometry, LocalGeometryMiller, default_miller_inputs
+from ..local_species import LocalSpecies
+from ..normalisation import SimulationNormalisation as Normalisation
+from ..normalisation import convert_dict, ureg
+from ..numerics import Numerics
+from ..file_utils import AbstractFileReader
+from ..templates import gk_templates
+from ..typing import PathLike
+from .gk_input import GKInput
+from .gk_output import Coords, Eigenvalues, Fields, Fluxes, GKOutput, Moments
+
+
+@GKInput.reader("GX")
+class GKInputGX(GKInput):
+    """
+    Class that can read GX input files, and produce
+    Numerics, LocalSpecies, and LocalGeometry objects
+    """
+
+    code_name = "GX"
+    default_file_name = "input.in"
+    norm_convention = "gx"
+
+    # gx does not calculate geometric information except for the simplest models
+    # Preferred method is to read geometric coefficients from a NetCDF file
+    # Secondary method is to read geometric coefficients from the GS2 eik.out file
+    pyro_gx_miller = {
+        "rho": ["Geometry", "rhoc"],
+        "Rmaj": ["Geometry", "rmaj"],
+        "q": ["Geometry", "qinp"],
+        "kappa": ["Geometry", "akappa"],
+        "shat": ["Geometry", "shat"],
+        "shift": ["Geometry", "shift"],
+        "beta_prime": ["Geometry", "beta_prime_input"],
+    }
+
+    pyro_gx_miller_defaults = {
+        "rho": 0.5,
+        "Rmaj": 1.0,
+        "q": 1.4,
+        "kappa": 1.0,
+        "shat": 0.8,
+        "shift": 0.0,
+        "beta_prime": 0.0,
+    }
+
+    pyro_gx_species = {
+        "mass": "mass",
+        "z": "z",
+        "dens": "dens",
+        "temp": "temp",
+        "nu": "vnewk",
+        "inverse_lt": "tprim",
+        "inverse_ln": "fprim",
+        "inverse_lv": "uprim",
+        "type": "type",
+    }
+
+    def read_from_file(self, filename: PathLike) -> Dict[str, Any]:
+        """
+        Reads GX input file into a dictionary
+        """
+        result = super().read_from_file(filename)
+        return result
+
+    def read_str(self, input_string: str) -> Dict[str, Any]:
+        """
+        Reads GX input file given as string
+        Uses default read_str, which assumes input_string has valid toml syntax
+        """
+        result = super().read_str(input_string)
+        return result
+
+    def read_dict(self, input_dict: dict) -> Dict[str, Any]:
+        """
+        Reads GX input file given as dict
+        Uses default read_dict, which assumes input is a dict
+        """
+        return super().read_dict(input_dict)
+
+    def verify_file_type(self, filename: PathLike):
+        """
+        Ensure this file is a valid gx input file, and that it contains sufficient
+        info for Pyrokinetics to work with
+        """
+        # The following keys are not strictly needed for a GX input file,
+        # but they are needed by Pyrokinetics
+        expected_keys = [
+            "Domain",
+            "Dimensions",
+            "Time",
+            "Initialization",
+            "Restart",
+            "Dissipation",
+            "Diagnostics",
+            "Boltzmann",
+            "Geometry",
+            "Physics",
+            "species",
+        ]
[…]
