+++ src/pyrokinetics/normalisation.py
@@ -248,6 +248,7 @@ NORMALISATION_CONVENTIONS = {
     "pyrokinetics": Convention("pyrokinetics"),
     "cgyro": Convention("cgyro", bref=ureg.bref_Bunit, rhoref=ureg.rhoref_unit),
     "gs2": Convention("gs2", vref=ureg.vref_most_probable, rhoref=ureg.rhoref_gs2),
+    "gx": Convention("gx", vref=ureg.vref_nrl, rhoref=ureg.rhoref_pyro),
     "gene": Convention("gene", lref=ureg.lref_major_radius, rhoref=ureg.rhoref_pyro),
     "imas": Convention(
         "imas",
