+++ src/pyrokinetics/pyro.py
@@ -18,6 +18,15 @@ import warnings
 import xarray as xr
 import numpy as np
 import f90nml
+try:
+    import tomllib
+except ModuleNotFoundError:
+    try:
+        import tomli as tomllib
+    except ModuleNotFoundError:
+        print("Please install tomli or upgrade to Python >= 3.11")
+        exit()
+
 
 from pathlib import Path
 from typing import Optional, List, Dict, Any, Union
@@ -69,20 +78,20 @@ class Pyro:
     eq_kwargs : Optional[Dict[str, Any]] = None
         Keyword arguments to be used when building an Equilibrium object
     kinetics_file : PathLike, default ``None``
-        Filename for outputs from a global kinetics code, such as SCENE, JETTO,
+        Filename for outputs from a global kinetics code, such as SCENE, JETTO, T3D,
         TRANSP, or pFile. When passed, this will set the 'kinetics' attribute. This can be used to
         set local kinetics using the function load_local_kinetics or load_local.
     kinetics_type : str, default ``None``
         Type of kinetics file. When set, this will skip file type inference. Possible
-        values are SCENE, JETTO, TRANSP, or pFile. If set to None, the file type is inferred
+        values are SCENE, JETTO, T3D, TRANSP, or pFile. If set to None, the file type is inferred
         automatically.
     kinetics_kwargs : Optional[Dict[str, Any]] = None
         Keyword arguments to be used when building a Kinetics object.
     gk_file : PathLike, default ``None``
-        Filename for a gyrokinetics input file (GS2, GENE, CGYRO). When passed, the
+        Filename for a gyrokinetics input file (GS2, GENE, CGYRO, GX). When passed, the
         attributes 'local_geometry', 'local_species', and 'numerics' are set.
     gk_output_file : PathLike, default ``None``
-        Filename or directory name for gyrokinetics output file(s). For GS2, the user
+        Filename or directory name for gyrokinetics output file(s). For GS2 or GX, the user
         should pass the '.out.nc' NetCDF4 file. For CGYRO, the user should pass the
         directory containing output files with their standard names. For GENE, the user
         should pass one of 'parameters_####', 'nrg_####' or 'field_####', where #### is
@@ -91,7 +100,7 @@ class Pyro:
         containing 'parameters_0000', 'field_0000' and 'nrg_0000'.
     gk_code : str, default ``None``
         Type of gyrokinetics input file and output file. When set, this will skip file
-        type inference. Possible values are 'GS2', 'CGYRO', or 'GENE'. If set to None,
+        type inference. Possible values are 'GS2', 'CGYRO', 'GX', or 'GENE'. If set to None,
         the file type is inferred automatically. If gk_code is set, but no gk_file is
         provided, the corresponding default template file will be read.
     gk_type : str, default ``None``
@@ -276,8 +285,8 @@ class Pyro:
     @property
     def supported_kinetics_types(self) -> List[str]:
         """
-        Returns a list of supported `Kinetics` types, expressed as strings (e.g. JETTO,
-        SCENE, TRANSP). The user can add new `Kinetics` reader classes by 'registering'
+        Returns a list of supported `Kinetics` types, expressed as strings (e.g. JETTO, T3D,
+        SCENE, T3D, TRANSP). The user can add new `Kinetics` reader classes by 'registering'
         them with `Kinetics.reader`
 
         Returns
@@ -294,7 +303,7 @@ class Pyro:
     def gk_code(self) -> Union[str, None]:
         """
         The current gyrokinetics context, expressed as a string. This is typically the
-        name of the gyrokinetics code (GS2, CGYRO, GENE, etc). If there is no
+        name of the gyrokinetics code (GS2, CGYRO, GENE, GX, etc). If there is no
[…]
