### 2023-10-12 6213e93
Bump pyrokinetics-plugin-examples version
