### 2023-10-12 7b8c9a8
Skip plugin tests if library is missing
Also reformatted test file with isort/docformatter
