+++ tests/test_plugins.py
@@ -1,32 +1,38 @@
 from pathlib import Path
 
 import pytest
-from pyrokinetics_plugin_examples import make_plugin_example_file
 
 from pyrokinetics.equilibrium import (
     Equilibrium,
-    supported_equilibrium_types,
     read_equilibrium,
+    supported_equilibrium_types,
 )
-from pyrokinetics.kinetics import Kinetics, supported_kinetics_types, read_kinetics
 from pyrokinetics.gk_code import (
     GKInput,
     GKOutput,
-    supported_gk_input_types,
-    supported_gk_output_types,
     read_gk_input,
     read_gk_output,
+    supported_gk_input_types,
+    supported_gk_output_types,
 )
+from pyrokinetics.kinetics import Kinetics, read_kinetics, supported_kinetics_types
 from pyrokinetics.normalisation import SimulationNormalisation
+
+# Skip tests if plugin library is missing
+try:
+    from pyrokinetics_plugin_examples import make_plugin_example_file
+except ImportError:
+    pytest.skip("Plugin test library not found.", allow_module_level=True)
 
 
 @pytest.fixture
 def plugin_file(tmp_path: Path) -> Path:
-    """
-    The pyrokinetics-plugin-examples repository contains a number of different
-    plugin implementations, including variants for Equilibrium, Kinetics,
-    GKInput, and GKOutput. However, all share the same input file. This fixture
-    creates a new copy of this universal input file.
+    """The pyrokinetics-plugin-examples repository contains a number of different plugin
+    implementations, including variants for Equilibrium, Kinetics, GKInput, and
+    GKOutput.
+
+    However, all share the same input file. This fixture creates a new copy of this
+    universal input file.
     """
     d = tmp_path / "plugin_test_files"
     d.mkdir(parents=True, exist_ok=True)
