### 2023-10-12 25fa1ea
Cleaned up the lint errors
