+++ src/pyrokinetics/gk_code/gx.py
@@ -3,12 +3,20 @@ from copy import copy
 from itertools import product
 from pathlib import Path
 from typing import Any, Dict, Optional, Tuple
+try:
+    import tomllib
+except ModuleNotFoundError:
+    try:
+        import tomli as tomllib
+    except ModuleNotFoundError:
+        print("Please install tomli or upgrade to Python >= 3.11")
+        exit()
 
 import numpy as np
 import xarray as xr
 from cleverdict import CleverDict
 
-from ..constants import pi, sqrt2
+from ..constants import pi
 from ..local_geometry import LocalGeometry, LocalGeometryMiller, default_miller_inputs
 from ..local_species import LocalSpecies
 from ..normalisation import SimulationNormalisation as Normalisation
@@ -71,7 +79,8 @@ class GKInputGX(GKInput):
         """
         Reads GX input file into a dictionary
         """
-        result = super().read_from_file(filename)
+        with open(filename, mode="rb") as f:
+            result = tomllib.load(f)
         return result
 
     def read_str(self, input_string: str) -> Dict[str, Any]:
@@ -123,7 +132,7 @@ class GKInputGX(GKInput):
 
     def is_nonlinear(self) -> bool:
         try:
-            is_nonlinear = self.data["Control"]["nonlinear_mode"] == True
+            is_nonlinear = self.data["Control"]["nonlinear_mode"]
             return is_nonlinear
         except KeyError:
             return False
@@ -136,7 +145,7 @@ class GKInputGX(GKInput):
 
     def get_local_geometry(self) -> LocalGeometry:
         """
-        Should return local geometry by delegating to more specific functions, but 
+        Should return local geometry by delegating to more specific functions, but
         instead, needs a file with pre-processed geometric data or a VMEC output file
         GX provides Python code to perform the pre-processing step
         """
@@ -237,12 +246,12 @@ class GKInputGX(GKInput):
         for i_sp in range(self.data["Dimensions"]["nspecies"]):
             species_data = CleverDict()
 
-            gx_key = f"species_parameters_{i_sp + 1}"
+            gx_key = f"species[{i_sp}]"
 
             gx_data = self.data[gx_key]
 
             for pyro_key, gx_key in self.pyro_gx_species.items():
-                species_data[pyro_key] = gs2_data[gx_key]
+                species_data[pyro_key] = gx_data[gx_key]
 
             species_data.vel = 0.0 * ureg.vref_nrl
             species_data.inverse_lv = 0.0 / ureg.lref_minor_radius
@@ -280,7 +289,7 @@ class GKInputGX(GKInput):
         #         self.data["parameters"]["zeff"] * ureg.elementary_charge
         #     )
         # else:
-        # I suppose this is the right thing to do? 
+        # I suppose this is the right thing to do?
         local_species.zeff = 1.0 * ureg.elementary_charge
 
         return local_species
@@ -314,24 +323,25 @@ class GKInputGX(GKInput):
         if box["jtwist"] == 0 and box["boundary"] == "periodic":
             raise RuntimeError("jtwist = 0 is not permitted with periodic boundary conditions")
 
-        if "jtwist" not in keys:
-            if "x0" in keys:
-                jtwist_default = min(round(2 * pi * abs(shat) * box["x0"]/box["y0"]), 1) 
-            elif            
-                jtwist_default = min(round(2 * pi * abs(shat)), 1)
-            
         shat_params = self.pyro_gx_miller["shat"]
         shat = self.data[shat_params[0]][shat_params[1]]
+
+        if "jtwist" not in keys:
+            if "x0" in keys:
[…]
