+++ src/pyrokinetics/gk_code/gx.py
@@ -3,6 +3,7 @@ from copy import copy
 from itertools import product
 from pathlib import Path
 from typing import Any, Dict, Optional, Tuple
+
 try:
     import tomllib
 except ModuleNotFoundError:
@@ -152,9 +153,7 @@ class GKInputGX(GKInput):
         gx_eq = self.data["Geometry"]["geo_option"]
 
         if gx_eq not in ["none", "slab", "const-curv", "s-alpha"]:
-            raise NotImplementedError(
-                f"GX equilibrium option {gx_eq} not implemented"
-            )
+            raise NotImplementedError(f"GX equilibrium option {gx_eq} not implemented")
 
         local_eq = self.data["Geometry"].get("local_eq", False)
         if not local_eq:
@@ -176,13 +175,11 @@ class GKInputGX(GKInput):
         # the midplane diameter to the Last Closed Flux Surface (LCFS) diameter
         if self.data["Geometry"]["bishop"] != 4:
             raise RuntimeError(
-                "Pyrokinetics requires GX input files to use "
-                "Geometry.bishop = 4"
+                "Pyrokinetics requires GX input files to use " "Geometry.bishop = 4"
             )
         if self.data["Geometry"]["irho"] != 2:
             raise RuntimeError(
-                "Pyrokinetics requires GX input files to use "
-                "Geometry.irho = 2"
+                "Pyrokinetics requires GX input files to use " "Geometry.irho = 2"
             )
 
         miller_data = default_miller_inputs()
@@ -194,15 +191,9 @@ class GKInputGX(GKInput):
 
         rho = miller_data["rho"]
         kappa = miller_data["kappa"]
-        miller_data["delta"] = np.sin(
-            self.data["Geometry"].get("tri", 0.0)
-        )
-        miller_data["s_kappa"] = (
-            self.data["Geometry"].get("akappri", 0.0) * rho / kappa
-        )
-        miller_data["s_delta"] = (
-            self.data["Geometry"].get("tripri", 0.0) * rho
-        )
+        miller_data["delta"] = np.sin(self.data["Geometry"].get("tri", 0.0))
+        miller_data["s_kappa"] = self.data["Geometry"].get("akappri", 0.0) * rho / kappa
+        miller_data["s_delta"] = self.data["Geometry"].get("tripri", 0.0) * rho
 
         # Get beta and beta_prime normalised to R_major(in case R_geo != R_major)
         r_geo = self.data["Geometry"].get("R_geo", miller_data["Rmaj"])
@@ -321,14 +312,18 @@ class GKInputGX(GKInput):
             raise RuntimeError("kx grid details not found in {keys}")
 
         if box["jtwist"] == 0 and box["boundary"] == "periodic":
-            raise RuntimeError("jtwist = 0 is not permitted with periodic boundary conditions")
+            raise RuntimeError(
+                "jtwist = 0 is not permitted with periodic boundary conditions"
+            )
 
         shat_params = self.pyro_gx_miller["shat"]
         shat = self.data[shat_params[0]][shat_params[1]]
 
         if "jtwist" not in keys:
             if "x0" in keys:
-                jtwist_default = min(round(2 * pi * abs(shat) * box["x0"]/box["y0"]), 1)
+                jtwist_default = min(
+                    round(2 * pi * abs(shat) * box["x0"] / box["y0"]), 1
+                )
             else:
                 jtwist_default = min(round(2 * pi * abs(shat)), 1)
 
@@ -338,7 +333,9 @@ class GKInputGX(GKInput):
         else:
             set_periodic = "periodic"
 
-        if ((set_periodic == "periodic") or (self.data["Domain"]["boundary"] == "periodic")) :
[…]
