### 2023-10-12 cfbab56
Add link to plugin examples to docs
