### 2023-10-12 88ff1bb
Field check for moment output and add time is no fields/moment read
