### 2023-10-12 b947912
Fix downsizing in CGYRO, add few more defaults
