+++ src/pyrokinetics/gk_code/cgyro.py
@@ -274,7 +274,7 @@ class GKInputCGYRO(GKInput, FileReader, file_type="CGYRO", reads=GKInput):
         # Assume pref*8pi*1e-7 = 1.0
         # FIXME Should not be modifying miller after creation
         ne_norm, Te_norm = self.get_ne_te_normalisation()
-        beta = self.data["BETAE_UNIT"] * ne_norm * Te_norm
+        beta = self.data.get("BETAE_UNIT", 0.0) * ne_norm * Te_norm
         if beta != 0:
             miller.B0 = 1 / (miller.bunit_over_b0 * beta**0.5)
         else:
@@ -323,7 +323,7 @@ class GKInputCGYRO(GKInput, FileReader, file_type="CGYRO", reads=GKInput):
         # Assume pref*8pi*1e-7 = 1.0
         # FIXME Should not be modifying mxh after creation
         ne_norm, Te_norm = self.get_ne_te_normalisation()
-        beta = self.data["BETAE_UNIT"] * ne_norm * Te_norm
+        beta = self.data.get("BETAE_UNIT", 0.0) * ne_norm * Te_norm
         if beta != 0:
             mxh.B0 = 1 / (mxh.bunit_over_b0 * beta**0.5)
         else:
@@ -363,7 +363,7 @@ class GKInputCGYRO(GKInput, FileReader, file_type="CGYRO", reads=GKInput):
         # FIXME Is this assumption general enough? Can't we get pref from local_species?
         # FIXME B0 = None can cause problems when writing
         ne_norm, Te_norm = self.get_ne_te_normalisation()
-        beta = self.data["BETAE_UNIT"] * ne_norm * Te_norm
+        beta = self.data.get("BETAE_UNIT", 0.0) * ne_norm * Te_norm
         if beta != 0:
             fourier.B0 = 1 / (fourier.bunit_over_b0 * beta**0.5)
         else:
@@ -411,7 +411,7 @@ class GKInputCGYRO(GKInput, FileReader, file_type="CGYRO", reads=GKInput):
             if species_data.z == -1:
                 name = "electron"
                 species_data.nu = (
-                    self.data["NU_EE"] * ureg.vref_nrl / ureg.lref_minor_radius
+                    self.data.get("NU_EE", 0.1) * ureg.vref_nrl / ureg.lref_minor_radius
                 )
             else:
                 ion_count += 1
@@ -465,7 +465,7 @@ class GKInputCGYRO(GKInput, FileReader, file_type="CGYRO", reads=GKInput):
 
         numerics_data = {}
 
-        nfields = self.data["N_FIELD"]
+        nfields = self.data.get("N_FIELD", 1)
 
         numerics_data["phi"] = nfields >= 1
         numerics_data["apar"] = nfields >= 2
@@ -495,7 +495,7 @@ class GKInputCGYRO(GKInput, FileReader, file_type="CGYRO", reads=GKInput):
 
         ne_norm, Te_norm = self.get_ne_te_normalisation()
         numerics_data["beta"] = (
-            self.data["BETAE_UNIT"] * ureg.beta_ref_ee_Bunit * ne_norm * Te_norm
+            self.data.get("BETAE_UNIT", 0.0) * ureg.beta_ref_ee_Bunit * ne_norm * Te_norm
         )
 
         numerics_data["gamma_exb"] = (
@@ -895,6 +895,13 @@ class GKOutputReaderCGYRO(FileReader, file_type="CGYRO", reads=GKOutput):
         # Process time data
         time = raw_data["time"][:, 0]
 
+        if  len(time) % downsize != 0:
+            residual = len(time) % downsize - downsize
+        else:
+            residual = 0
+
+        time = time[::downsize]
+
         # Process grid data
         grid_data = raw_data["grids"]
         nky = int(grid_data[0])
[…]
