### 2023-10-16 83aeee2
Add test for downsample
