### 2023-10-16 6651a72
add note on TGLF collision frequency
