### 2023-10-16 e98840d
Switch to using only angular velocity rather than the toroidal velocity
