+++ src/pyrokinetics/gk_code/gs2.py
@@ -59,7 +59,6 @@ class GKInputGS2(GKInput, FileReader, file_type="GS2", reads=GKInput):
         "nu": "vnewk",
         "inverse_lt": "tprim",
         "inverse_ln": "fprim",
-        "inverse_lv": "uprim",
     }
 
     def read_from_file(self, filename: PathLike) -> Dict[str, Any]:
@@ -224,12 +223,6 @@ class GKInputGS2(GKInput, FileReader, file_type="GS2", reads=GKInput):
 
         ne_norm, Te_norm = self.get_ne_te_normalisation()
 
-        domega_drho = (
-            self.data["theta_grid_parameters"]["qinp"]
-            / self.data["theta_grid_parameters"]["rhoc"]
-            * self.data["dist_fn_knobs"].get("g_exb", 0.0)
-        )
-
         # Load each species into a dictionary
         for i_sp in range(self.data["species_knobs"]["nspec"]):
             species_data = CleverDict()
@@ -241,10 +234,17 @@ class GKInputGS2(GKInput, FileReader, file_type="GS2", reads=GKInput):
             for pyro_key, gs2_key in self.pyro_gs2_species.items():
                 species_data[pyro_key] = gs2_data[gs2_key]
 
-            species_data.vel = 0.0 * ureg.vref_most_probable
-            species_data.inverse_lv = 0.0 / ureg.lref_minor_radius
+            domega_drho = (
+                gs2_data.get("uprim", 0.0)
+                / self.data["theta_grid_parameters"]["rmaj"]
+            )
+            species_data.omega0 = (
+                self.data["dist_fn_knobs"].get("mach", 0.0)
+                * ureg.vref_most_probable
+                / ureg.lref_minor_radius
+            )
             species_data.domega_drho = (
-                domega_drho * ureg.vref_most_probable / ureg.lref_minor_radius**2
+                domega_drho * ureg.vref_most_probable / ureg.lref_minor_radius ** 2
             )
 
             if species_data.z == -1:
@@ -261,8 +261,8 @@ class GKInputGS2(GKInput, FileReader, file_type="GS2", reads=GKInput):
             species_data.nu *= ureg.vref_most_probable / ureg.lref_minor_radius
             species_data.temp *= ureg.tref_electron / Te_norm
             species_data.z *= ureg.elementary_charge
-            species_data.inverse_lt *= ureg.lref_minor_radius**-1
-            species_data.inverse_ln *= ureg.lref_minor_radius**-1
+            species_data.inverse_lt *= ureg.lref_minor_radius ** -1
+            species_data.inverse_ln *= ureg.lref_minor_radius ** -1
 
             # Add individual species data to dictionary of species
             local_species.add_species(name=name, species_data=species_data)
@@ -507,6 +507,12 @@ class GKInputGS2(GKInput, FileReader, file_type="GS2", reads=GKInput):
                     local_norm.gs2
                 )
 
+            self.data[species_key]["uprim"] = (
+                local_species[name]["domega_drho"]
+                / self.data["theta_grid_parameters"]["rmaj"]
+            )
+
+        self.data["dist_fn_knobs"]["mach"] = local_species.electron.omega0
         self.data["knobs"]["zeff"] = local_species.zeff
 
         beta_ref = local_norm.gs2.beta if local_norm else 0.0
