+++ src/pyrokinetics/gk_code/cgyro.py
@@ -276,7 +276,7 @@ class GKInputCGYRO(GKInput, FileReader, file_type="CGYRO", reads=GKInput):
         ne_norm, Te_norm = self.get_ne_te_normalisation()
         beta = self.data["BETAE_UNIT"] * ne_norm * Te_norm
         if beta != 0:
-            miller.B0 = 1 / (miller.bunit_over_b0 * beta ** 0.5)
+            miller.B0 = 1 / (miller.bunit_over_b0 * beta**0.5)
         else:
             miller.B0 = None
 
@@ -286,7 +286,7 @@ class GKInputCGYRO(GKInput, FileReader, file_type="CGYRO", reads=GKInput):
 
         if miller.B0 is not None:
             miller.beta_prime = (
-                -local_species.inverse_lp.m * beta_prime_scale / miller.B0 ** 2
+                -local_species.inverse_lp.m * beta_prime_scale / miller.B0**2
             )
         else:
             miller.beta_prime = 0.0
@@ -325,7 +325,7 @@ class GKInputCGYRO(GKInput, FileReader, file_type="CGYRO", reads=GKInput):
         ne_norm, Te_norm = self.get_ne_te_normalisation()
         beta = self.data["BETAE_UNIT"] * ne_norm * Te_norm
         if beta != 0:
-            mxh.B0 = 1 / (mxh.bunit_over_b0 * beta ** 0.5)
+            mxh.B0 = 1 / (mxh.bunit_over_b0 * beta**0.5)
         else:
             mxh.B0 = None
 
@@ -335,7 +335,7 @@ class GKInputCGYRO(GKInput, FileReader, file_type="CGYRO", reads=GKInput):
 
         if mxh.B0 is not None:
             mxh.beta_prime = (
-                -local_species.inverse_lp.m * beta_prime_scale / mxh.B0 ** 2
+                -local_species.inverse_lp.m * beta_prime_scale / mxh.B0**2
             )
         else:
             mxh.beta_prime = 0.0
@@ -365,7 +365,7 @@ class GKInputCGYRO(GKInput, FileReader, file_type="CGYRO", reads=GKInput):
         ne_norm, Te_norm = self.get_ne_te_normalisation()
         beta = self.data["BETAE_UNIT"] * ne_norm * Te_norm
         if beta != 0:
-            fourier.B0 = 1 / (fourier.bunit_over_b0 * beta ** 0.5)
+            fourier.B0 = 1 / (fourier.bunit_over_b0 * beta**0.5)
         else:
             fourier.B0 = None
 
@@ -375,7 +375,7 @@ class GKInputCGYRO(GKInput, FileReader, file_type="CGYRO", reads=GKInput):
 
         if fourier.B0 is not None:
             fourier.beta_prime = (
-                -local_species.inverse_lp.m * beta_prime_scale / fourier.B0 ** 2
+                -local_species.inverse_lp.m * beta_prime_scale / fourier.B0**2
             )
         else:
             fourier.beta_prime = 0.0
@@ -406,7 +406,7 @@ class GKInputCGYRO(GKInput, FileReader, file_type="CGYRO", reads=GKInput):
                 self.data.get("MACH", 0.0) * ureg.vref_nrl / ureg.lref_minor_radius
             )
             species_data.domega_drho = (
-                domega_drho * ureg.vref_nrl / ureg.lref_minor_radius ** 2
+                domega_drho * ureg.vref_nrl / ureg.lref_minor_radius**2
             )
 
             if species_data.z == -1:
@@ -425,8 +425,8 @@ class GKInputCGYRO(GKInput, FileReader, file_type="CGYRO", reads=GKInput):
             species_data.mass *= ureg.mref_deuterium
[…]
