+++ src/pyrokinetics/gk_code/cgyro.py
@@ -403,8 +403,10 @@ class GKInputCGYRO(GKInput, FileReader, file_type="CGYRO", reads=GKInput):
                 species_data[p_key] = self.data[c_key]
 
             species_data.omega0 = (
-                self.data.get("MACH", 0.0) * ureg.vref_nrl / ureg.lref_minor_radius / self.data["RMAJ"]
-
+                self.data.get("MACH", 0.0)
+                * ureg.vref_nrl
+                / ureg.lref_minor_radius
+                / self.data["RMAJ"]
             )
             species_data.domega_drho = (
                 domega_drho * ureg.vref_nrl / ureg.lref_minor_radius**2
