### 2023-10-17 a31f861
Remove uprim from GS2
