### 2023-10-17 15bf500
Remove potential divide by 0
