### 2023-10-17 1bd256b
Parametrize test and update gene template
