### 2023-10-17 64ec746
Parametrize test and update gene template
