### 2023-10-17 2a7f939
Add field check for NL back in
