### 2023-10-17 #5866 @rogeriojorge
This is the start of the review of the code **Pyrokinetics - A Python library to standardise gyrokinetic analysis**

Below, I describe the process of download, installation and running of Pytokinetics.

- Downloading from source:
No issues. However, I have to reference the fact that I had to download 118 MB of data which I believe comes from TRANSP files. Are those really necessary?

- Doing pip install -e .[tests] led me to the error:
no matches found: .[tests]
Installed normally using pip install -e .
Then installed from source with no issues

- Doing pytest --cov . led me to the error:
pytest: error: unrecognized arguments: --cov
had to install pytest-cov
Then ran the test.

This outputted the 89% code coverage and 3 failed tests

====================================================== short test summary info ======================================================
FAILED tests/local_geometry/test_fourier.py::test_load_from_eq - AssertionError: cN difference: [2.54681480e-08 1.21464990e-07 5.16698228e-07 1.06052750e-06
FAILED tests/local_geometry/test_fourier_cgyro.py::test_load_from_eq - AssertionError: aR difference: [8.46300070e-08 8.31583997e-07 2.86840035e-06 6.25578335e-06
FAILED tests/local_geometry/test_mxh.py::test_load_from_eq - AssertionError: bunit_over_b0 difference: 0.0004917421190069327
================================= 3 failed, 778 passed, 1 skipped, 268 warnings in 85.19s (0:01:25) =================================

While the code coverage seems good, I am not able to assess if not passing the tests is significant or not. **Can the authors address this?**

- Code running
I then proceeded with the README and ran
pyro = Pyro(gk_file="gs2Input-linear.in")
for the file in the link [https://raw.githubusercontent.com/rogeriojorge/microstability_optimization/main/src/GK_inputs/gs2Input-linear.in](https://raw.githubusercontent.com/rogeriojorge/microstability_optimization/main/src/GK_inputs/gs2Input-linear.in). I got the error below. **Can the authors address this issue?**

- How to mitigate these issues
I urge the authors to add a Docker file to Pyrokinetics. It will make newcomers like me to be able to use the terminal interface much more efficiently and use it in a much more efficient manner.
**Can the authors address this?**

Once these three topics are addressed, I will continue with the remaining check points.

========================
--------------------------------

Traceback (most recent call last):
  File "/Users/rogeriojorge/local/pyrokinetics/src/pyrokinetics/factory.py", line 37, in type
    return self._registered_types[key]
           ~~~~~~~~~~~~~~~~~~~~~~^^^^^
KeyError: 'gs2Input-linear.in'

Traceback (most recent call last):
  File "/Users/rogeriojorge/local/pyrokinetics/src/pyrokinetics/file_utils.py", line 128, in type
    return super().type(key)
           ^^^^^^^^^^^^^^^^^
  File "/Users/rogeriojorge/local/pyrokinetics/src/pyrokinetics/factory.py", line 39, in type
    raise KeyError(
KeyError: "'gs2Input-linear.in' is not registered with FileReader factory"

Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/rogeriojorge/local/pyrokinetics/src/pyrokinetics/pyro.py", line 174, in __init__
    self.read_gk_file(gk_file, gk_code)
  File "/Users/rogeriojorge/local/pyrokinetics/src/pyrokinetics/pyro.py", line 852, in read_gk_file
    gk_input = read_gk_input(gk_file, file_type=gk_code)
               ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/rogeriojorge/local/pyrokinetics/src/pyrokinetics/gk_code/gk_input.py", line 221, in read_gk_input
    gk_input = GKInput._factory(file_type if file_type is not None else path)
               ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/rogeriojorge/local/pyrokinetics/src/pyrokinetics/factory.py", line 66, in __call__
    return self.create(key, *args, **kwargs)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/rogeriojorge/local/pyrokinetics/src/pyrokinetics/factory.py", line 45, in create
    return self.type(key)(*args, **kwargs)
           ^^^^^^^^^^^^^^
  File "/Users/rogeriojorge/local/pyrokinetics/src/pyrokinetics/file_utils.py", line 130, in type
    return super().type(self._infer_file_type(key))
                        ^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/rogeriojorge/local/pyrokinetics/src/pyrokinetics/file_utils.py", line 145, in _infer_file_type
    raise KeyError(err_msg)
