### 2023-10-17 bpatel2107/pyrokinetics#262 @bpatel2107
JOSS paper review - Reviewer 1 
The JOSS paper review has started and there are some issues that one of the reviewers has come across. Opening this issue to work through the issues faced by the first reviewer.

https://github.com/openjournals/joss-reviews/issues/5866#issuecomment-1766842546
