### 2023-10-17 bpatel2107/pyrokinetics#262 @LiamPattinson
- Not sure what's going on with the `pip install -e .[tests]` error. Maybe it's just not running on the latest pip? It may be worth adding `pip install --upgrade pip` throughout the README file.
- The failing tests are the same few that keep falling over in our CI. I think these golden answer tests may need to be replaced, as they seem far too sensitive to the user's system (or are they indicative that the Fourier geometry fitting actually isn't working?).
- The GS2 file that won't read seems to be because it's using features we don't (yet?) support. The error message reported is `KeyError: "'my_gs2.in' is neither a registered key nor a valid file."`, which isn't helpful. If you try running `p = Pyro(gk_file="my_gs2.in", gk_code="GS2")`, you get the much more helpful `NotImplementedError: GS2 equilibrium option file not implemented`. We should add some code to catch the generic KeyError that was raised, explain that the cause is an invalid input file, and recommend specifying `gk_code` to get a more useful error message. We can do the same for equilibrium and kinetics too. Unfortunately, I can't think of a way to get the correct error message without adding `gk_code`/`eq_type`/etc, as in the process of figuring out the file type, every invalid file reader raises their own exception, and there's no way to tell which of these exceptions are relevant.
- I can add a Dockerfile, and I can make it so that a new one is published automatically on each release. I don't have repository access rights to add workflow secrets though, so I'll have to coordinate that with somebody who does.
