### 2023-10-18 bbe94af
Write normalisations for GENE/GS2
