### 2023-10-18 b7d73c2
Write normalisations for GENE/GS2
