+++ src/pyrokinetics/normalisation.py
@@ -628,7 +628,9 @@ class SimulationNormalisation(Normalisation):
 
         if lref_minor_radius and lref_major_radius:
             if lref_major_radius != pyro.local_geometry.Rmaj * lref_minor_radius:
-                raise ValueError("Specified major radius and minor radius do not match, please check the data")
+                raise ValueError(
+                    "Specified major radius and minor radius do not match, please check the data"
+                )
         elif lref_minor_radius:
             lref_major_radius = lref_minor_radius / pyro.local_geometry.Rmaj
         elif lref_major_radius:
