### 2023-10-18 0a0ccb5
Set rhoref more rigorously and fix tests
