+++ src/pyrokinetics/gk_code/cgyro.py
@@ -599,14 +599,13 @@ class GKInputCGYRO(GKInput, FileReader, file_type="CGYRO", reads=GKInput):
 
         for i_sp, name in enumerate(local_species.names):
             pyro_cgyro_species = self.get_pyro_cgyro_species(i_sp + 1)
-
             for pyro_key, cgyro_key in pyro_cgyro_species.items():
                 self.data[cgyro_key] = local_species[name][pyro_key].to(
                     local_norm.cgyro
                 )
         self.data["MACH"] = local_species.electron.omega0 * self.data["RMAJ"]
         self.data["GAMMA_P"] = (
-            -local_species.electron.domega_drho
+            -local_species.electron.domega_drho.to(local_norm.cgyro)
             * self.data["RMAJ"]
             * local_norm.cgyro.lref
         )
