### 2023-10-18 902cad2
Set rhoref more rigorously and fix tests
