### 2023-10-18 f55b956
Set gamma_exb only when needed
