### 2023-10-18 97b0848
Add GX to factory reader
