### 2023-10-18 3cf74f9
Return input file as self.data when reading GX files
