### 2023-10-18 158a0db
Set up GACODE kinetics/equilibrium
