+++ src/pyrokinetics/equilibrium/__init__.py
@@ -9,6 +9,7 @@ from .flux_surface import FluxSurface
 # Import each built-in reader to register them with Equilibrium
 from . import geqdsk  # noqa
 from . import transp  # noqa
+from . import gacode  # noqa
 
 # Register external plugins with Equilibrium
 from ..plugins import register_file_reader_plugins
