### 2023-10-18 d84a4e4
Set up GACODE kinetics/equilibrium
