+++ src/pyrokinetics/equilibrium/gacode.py
@@ -1,0 +1,210 @@
+from typing import Optional
+
+# Can't use xarray, as TRANSP has a variable called X which itself has a dimension
+# called X
+import numpy as np
+from scipy.interpolate import RBFInterpolator
+from pygacode import expro
+
+from ..file_utils import FileReader
+from ..typing import PathLike
+from ..units import ureg as units, UnitSpline
+from .equilibrium import Equilibrium
+
+
+class EquilibriumReaderTRANSP(FileReader, file_type="GACODE", reads=Equilibrium):
+    r"""
+    Class that can read input.gacode files. Rather than creating instances of this
+    class directly, users are recommended to use the function `read_equilibrium`.
+
+    See Also
+    --------
+    Equilibrium: Class representing a global tokamak equilibrium.
+    read_equilibrium: Read an equilibrium file, return an `Equilibrium`.
+    """
+
+    def read_from_file(
+        self,
+        filename: PathLike,
+        nR: Optional[int] = None,
+        nZ: Optional[int] = None,
+        clockwise_phi: bool = True,
+        cocos: Optional[int] = None,
+        neighbors: Optional[int] = 64,
+    ) -> Equilibrium:
+        r"""
+        Read in TRANSP netCDF and creates Equilibrium object.
+
+        TRANSP makes use of radial grids, and these are interpolated onto a Cartesian
+        RZ grid. Additional keyword-only arguments may be provided to control the
+        resolution of the Cartesian grid, and to choose the time at which the
+        equilibrium is taken.
+
+        Parameters
+        ----------
+        filename: PathLike
+            Path to the TRANSP netCDF file.
+        nR: Optional[int]
+            The number of grid points in the major radius direction. By default, this
+            is set to the number of radial grid points in the TRANSP file.
+        nZ: Optional[int]
+            The number of grid points in the vertical direction. By default, this
+            is set to the number of radial grid points in the TRANSP file.
+        clockwise_phi: bool, default False
+            Determines whether the :math:`\phi` grid increases clockwise or
+            anti-clockwise when viewed from above. Used to determine COCOS convention of
+            the inputs.
+        cocos: Optional[int]
+            If set, asserts that the GEQDSK file follows that COCOS convention, and
+            neither ``clockwise_phi`` nor the file contents will be used to identify
+            the actual convention in use. The resulting Equilibrium is always converted
+            to COCOS 11.
+        neighbors: Optional[int]
+            Sets number of nearest neighbours to use when performing the interpolation
+            to flux surfaces to R,Z. By default, this is 16
+
+        Raises
+        ------
+        FileNotFoundError
+            If ``filename`` is not a valid file.
+        RuntimeError
+            If the user provides both ``time`` and ``time_index``.
+        ValueError
+            If nr or nz are negative.
+        IndexError
+            If ``time_index`` is provided, but is out of bounds.
+
+        Returns
+        -------
+        Equilibrium
+        """
+        # Define some units to be used later
+        # Note that length units are in centimeters!
+        # This is not consistent throughout. Pressure is in Pascal as usual, not
+        # Newtons per centimeter^2. However, it does affect our units for F.
+        len_units = units.meter
+        psi_units = units.weber / units.radian
+
+        # Open data file, get generic data
+        expro.expro_read(filename, 0)
+
[…]
