### 2023-10-19 777bbcd
Add test for gacode (fix bug in other tests)
