### 2023-10-20 dba1141
Option to calculate omega from Er in JETTO
