### 2023-10-25 bpatel2107/pyrokinetics#262 @ZedThree
Personally, I think having a Docker container for pyrokinetics is little bit of overkill, and it would be much better to sort out any installation issues -- this will also make it easier to install in a container if we really need to go down that route. We might instead also check that the CLI works properly with `pipx`.

I do think it is probably a good idea to move most of the data files that we only need for the tests out to either another repo or something like zenodo.

Then it looks like the other issues are probably just making nicer error messages.

And then the other thing we might want to think about it is sorting out the dependencies, either making some optional, or doing lazy importing to make start up time much faster.
