### 2023-10-25 #5866 @the-rccg
**Examples**: Currently I can only find one Example in the Documentation (https://pyrokinetics.readthedocs.io/en/latest/examples/index.html) that is a trivial two-liner:
```python
import pyrokinetics as pk
pyro = pk.Pyro(gk_file="gs2.input")
```
Considering the stated usecase to standardize input and output files, it feels like one example for each showcasing how to load and save should be the bare minimum. Ideally, some real world example should be done with the resulting object to help people get started with what could be done within the framework.
