### 2023-10-26 574c26a
fix typo in use_ave_ion_grid
