+++ tests/diagnostics/test_tglf.py
@@ -5,10 +5,10 @@ from numpy.testing import assert_allclose
 from pyrokinetics import template_dir
 from pyrokinetics.diagnostics import sum_ky_spectrum, get_sat_params
 
-template_path = os.path.join(template_dir, 'outputs/TGLF_transport/')
+template_path = os.path.join(template_dir, "outputs/TGLF_transport/")
 
 # Read number of fields, species, modes, ky, and moments
-with open(os.path.join(template_path, 'out.tglf.QL_flux_spectrum'), 'r') as f:
+with open(os.path.join(template_path, "out.tglf.QL_flux_spectrum"), "r") as f:
     lines = f.readlines()
 ntype, nspecies, nfield, nky, nmodes = list(map(int, lines[3].strip().split()))
 
@@ -25,17 +25,23 @@ QLw = np.array(ql).reshape(nspecies, nfield, nmodes, nky, ntype)
 # Write QL data into Xarray and reshape
 QL_data_array = xr.DataArray(
     data=QLw,
-    dims=["species", 'field', 'mode', 'ky', "type"],
+    dims=["species", "field", "mode", "ky", "type"],
     coords={
         "species": np.arange(nspecies),
         "field": np.arange(nfield),
         "mode": np.arange(nmodes),
         "ky": np.arange(nky),
-        "type": ['particle', 'energy', 'toroidal stress', 'parallel stress', 'exchange'],
+        "type": [
+            "particle",
+            "energy",
+            "toroidal stress",
+            "parallel stress",
+            "exchange",
+        ],
     },
 )
 
-QL_data = QL_data_array.transpose('ky', 'mode', 'species', 'field', 'type').data
+QL_data = QL_data_array.transpose("ky", "mode", "species", "field", "type").data
 particle_QL = QL_data[:, :, :, :, 0]
 energy_QL = QL_data[:, :, :, :, 1]
 toroidal_stress_QL = QL_data[:, :, :, :, 2]
@@ -43,52 +49,70 @@ parallel_stress_QL = QL_data[:, :, :, :, 3]
 exchange_QL = QL_data[:, :, :, :, 4]
 
 # Read spectral shift and ave_p0 (only needed for SAT0)
-with open(os.path.join(template_path, 'out.tglf.spectral_shift_spectrum'), 'r') as f:
+with open(os.path.join(template_path, "out.tglf.spectral_shift_spectrum"), "r") as f:
     kx0_e = np.loadtxt(f, skiprows=5, unpack=True)
-with open(os.path.join(template_path, 'out.tglf.ave_p0_spectrum'), 'r') as f:
+with open(os.path.join(template_path, "out.tglf.ave_p0_spectrum"), "r") as f:
     ave_p0 = np.loadtxt(f, skiprows=3, unpack=True)
 
 # Read scalar saturation parameters
 inputs = {}
-with open(os.path.join(template_path, 'out.tglf.scalar_saturation_parameters'), 'r') as f:
+with open(
+    os.path.join(template_path, "out.tglf.scalar_saturation_parameters"), "r"
+) as f:
     content = f.readlines()
     for line in content[1:]:
-        line = line.strip().split('\n')
-        if any([x.startswith(('!', 'UNITS', 'SAT_RULE', 'XNU_MODEL', 'ETG_FACTOR', 'R_unit', 'ALPHA_ZF', 'RULE')) for x in line]):
-            if line[0].startswith('R_unit'):
-                line = line[0].split(' = ')
+        line = line.strip().split("\n")
+        if any(
[…]
