### 2023-10-26 #5866 @the-rccg
Thank you for replying this fast! The How-To's seem to include examples and cover all important use cases I would wish to see. This then seems to simply be an issue of highlighting the fact and removing the duplicate examples heading. Maybe call it "How-To Examples"?

While the content is great, the presentation is still lacking. It took two clicks to find an error with wrong indentation (https://pyrokinetics.readthedocs.io/en/latest/howtos/linear_outputs.html). I would recommend going through the How-To's once more and maybe even including them into your test suite to ensure new users will not encounter Problems right away.

These How-to's also mix the CLI and Code interface; the previous example only uses the CLI (https://pyrokinetics.readthedocs.io/en/latest/howtos/convert.html). Just a small note that it would be beneficial to probably showcase both methods if both are possible within the code.

I'll mark the examples as sufficient, but would really encourage you to go through them once more to remove these inconsistencies.
