### 2023-10-31 83e2e18
Better error messages from file readers
