### 2023-10-31 42ec7b8
Better error messages from file readers
