### 2023-11-01 6b475bf
Update GS2 reader for next
