+++ src/pyrokinetics/gk_code/gs2.py
@@ -712,9 +712,13 @@ class GKOutputReaderGS2(FileReader, file_type="GS2", reads=GKOutput):
             # Old diagnostics (and eventually the single merged diagnostics)
             # input file stored as array of bytes
             if isinstance(input_file.data[0], np.ndarray):
-                input_str = "\n".join(("".join(np.char.decode(line)).strip() for line in input_file.data))
+                input_str = "\n".join(
+                    ("".join(np.char.decode(line)).strip() for line in input_file.data)
+                )
             else:
-                input_str = "\n".join((line.decode("utf-8") for line in input_file.data))
+                input_str = "\n".join(
+                    (line.decode("utf-8") for line in input_file.data)
+                )
         gk_input = GKInputGS2()
         gk_input.read_str(input_str)
         return raw_data, gk_input, input_str
