### 2023-11-02 ee41e9e
Remove pygacode from doc build
