+++ docs/conf.py
@@ -46,6 +46,9 @@ autosummary_generate = True
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
 
+# Mock imports that can't be installed
+autodoc_mock_imports = ["pygacode"]
+
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
