### 2023-11-02 f836494
Move autdoc_mock to later in file
