+++ docs/conf.py
@@ -46,9 +46,6 @@ autosummary_generate = True
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
 
-# Mock imports that can't be installed
-autodoc_mock_imports = ["pygacode"]
-
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
@@ -72,6 +69,9 @@ autodoc_default_options = {
 }
 autodoc_typehints = "description"
 autodoc_class_signature = "mixed"
+
+# Mock imports that can't be installed
+autodoc_mock_imports = ["pygacode"]
 
 # The default role for text marked up `like this`
 default_role = "any"
