+++ pyproject.toml
@@ -50,6 +50,7 @@ dependencies = [
 
 [project.optional-dependencies]
 docs = [
+    "sphinx-autoapi>= 3.0.0",
     "sphinx >= 5.3",
     "sphinx_autodoc_typehints >= 1.19",
     "sphinx-book-theme >= 0.4.0rc1",
