+++ .readthedocs.yaml
@@ -23,6 +23,10 @@ sphinx:
 # formats:
 #    - pdf
 
+conda:
+  enironment: .environment.yaml
+
+
 # Optionally declare the Python requirements required to build your docs
 python:
   install:
