+++ .readthedocs.yaml
@@ -9,7 +9,8 @@ version: 2
 build:
   os: ubuntu-20.04
   tools:
-    python: "3.9"
+    python: "mambaforge-22.9"
+
     # You can also specify other tool versions:
     # nodejs: "16"
     # rust: "1.55"
