+++ src/pyrokinetics/local_geometry/miller.py
@@ -1,9 +1,10 @@
-import numpy as np
 from typing import Tuple
+
+import numpy as np
 from scipy.optimize import least_squares  # type: ignore
-from .local_geometry import LocalGeometry
+
 from ..typing import ArrayLike
-from .local_geometry import default_inputs
+from .local_geometry import LocalGeometry, default_inputs
 
 
 def default_miller_inputs():
