+++ src/pyrokinetics/gk_code/__init__.py
@@ -1,11 +1,10 @@
 from platform import python_version_tuple
 
-from .gk_input import GKInput, read_gk_input, supported_gk_input_types
-from .gk_output import GKOutput, read_gk_output, supported_gk_output_types
-
 # Import built-in input and output readers to register them with GkInput and GkOutput.
 from .cgyro import GKInputCGYRO, GKOutputReaderCGYRO  # noqa
 from .gene import GKInputGENE, GKOutputReaderGENE  # noqa
+from .gk_input import GKInput, read_gk_input, supported_gk_input_types
+from .gk_output import GKOutput, read_gk_output, supported_gk_output_types
 from .gs2 import GKInputGS2, GKOutputReaderGS2  # noqa
 from .tglf import GKInputTGLF, GKOutputReaderTGLF  # noqa
 
