### 2023-11-03 96f2b51
Revert pygacode move to optional dependencies
