### 2023-11-03 4fc71e5
Revert pygacode move to optional dependencies
