+++ src/pyrokinetics/equilibrium/__init__.py
@@ -1,4 +1,5 @@
 from ..plugins import register_file_reader_plugins
+from . import gacode  # noqa
 from . import geqdsk  # noqa
 from . import transp  # noqa
 from .equilibrium import (
@@ -8,18 +9,6 @@ from .equilibrium import (
     supported_equilibrium_types,
 )
 from .flux_surface import FluxSurface
-
-pygacode_found: bool
-try:
-    import pygacode
-
-    pygacode_found = True
-    del pygacode
-except ImportError:
-    pygacode_found = False
-
-if pygacode_found:
-    from . import gacode  # noqa
 
 register_file_reader_plugins("equilibrium", Equilibrium)
 
