+++ pyproject.toml
@@ -45,12 +45,10 @@ dependencies = [
     "simplejson >= 3.17.6",
     "gitpython >= 3.1",
     "jetto-tools",
+    "pygacode",
 ]
 
 [project.optional-dependencies]
-gacode = [
-  "pygacode",
-]
 docs = [
     "sphinx >= 5.3",
     "sphinx_autodoc_typehints >= 1.19",
