+++ src/pyrokinetics/kinetics/__init__.py
@@ -1,10 +1,15 @@
-from ..plugins import register_file_reader_plugins
-from .gacode import KineticsReaderGACODE
-from .jetto import KineticsReaderJETTO  # noqa
+# isort: skip_file
+
 from .kinetics import Kinetics, read_kinetics, supported_kinetics_types
-from .pfile import KineticsReaderpFile  # noqa
+
+# Import each module to register the associated readers
 from .scene import KineticsReaderSCENE  # noqa
 from .transp import KineticsReaderTRANSP  # noqa
+from .pfile import KineticsReaderpFile  # noqa
+from .jetto import KineticsReaderJETTO  # noqa
+from .gacode import KineticsReaderGACODE  # noqa
+
+from ..plugins import register_file_reader_plugins
 
 register_file_reader_plugins("kinetics", Kinetics)
 
