### 2023-11-06 54f1d7d
Allow for multiple ky in linear runs and update tests for new growth_rate_tolerance value
