+++ tests/gk_code/test_gk_output_reader_cgyro.py
@@ -79,7 +79,7 @@ def test_infer_path_from_input_file_cgyro():
 # Update: Commit d3da468c accounts for new gkoutput structure
 # This data was gathered from templates/outputs/CGYRO_linear
 
-reference_data_commit_hash = "d3da468c"
+reference_data_commit_hash = "98122aed"
 
 
 @pytest.fixture(scope="class")
@@ -127,6 +127,7 @@ class TestCGYROGoldenAnswers:
             "eigenfunctions",
             "growth_rate",
             "mode_frequency",
+            "growth_rate_tolerance",
         ],
     )
     def test_data_vars(self, array_similar, var):
@@ -140,7 +141,6 @@ class TestCGYROGoldenAnswers:
             "input_file",
             "attribute_units",
             "title",
-            "growth_rate_tolerance",
         ],
     )
     def test_data_attrs(self, attr):
