### 2023-11-06 f9ed414
Allow for multiple ky in linear runs and update tests for new growth_rate_tolerance value
