+++ tests/gk_code/test_gk_output_reader_tglf.py
@@ -83,11 +83,9 @@ def test_read_tglf_transport():
 
 
 # Golden answer tests
-# Compares against results obtained using GKCode methods from commit 7d551eaa
-# Update: Commit d3da468c accounts for new gkoutput structure
 # This data was gathered from templates/outputs/TGLF_linear
 
-reference_data_commit_hash = "98122aed"
+reference_data_commit_hash = "54f1d7d1"
 
 
 @pytest.fixture(scope="class")
