### 2023-11-09 d3e53fb
Correct gs2 range defaults
