### 2023-11-13 #5866 @the-rccg
@bpatel2107 Last comment on documentation, I do not see any information on contribution guidelines and hence cannot check off the following bullet point. Would be great to see that added!
As an aside on documentation, the ReadMe does define the unit normalization convention readable for plasma physicists. It might be useful to have at least a glossary/explanation of terms for didactic purposes.
