### 2023-11-16 bf7ee9c
Calculate k_perp from metric terms
