### 2023-11-16 7407212
Update GS2 output to include k_perp
