+++ src/pyrokinetics/templates/outputs/GS2_linear/gs2.in
@@ -122,7 +122,7 @@
     write_phi_over_time = .true.
     write_apar_over_time = .true.
     write_bpar_over_time = .true.
-    write_nl_flux = .true.
+    write_fluxes = .true.
     nwrite = 50
     navg = 50
     omegatol = 0.0001
