### 2023-11-16 e3060b6
Update GS2 output to include k_perp
