### 2023-11-21 48cddae
Update k_perp to calculate in flux tube space
