### 2023-11-21 74aaa3b
Update ntheta=64 for gs2 template, drop tol and calculate theta_range
