### 2023-11-21 469f2be
Update ntheta=64 for gs2 template, drop tol and calculate theta_range
