### 2023-11-22 63e99fb
Update gacode.py to correct typo in docstring
