+++ docs/examples/read_output.rst
@@ -1,0 +1,13 @@
+===============
+Read GK Outputs
+===============
+
+
+Once a gyrokinetic simulation is complete it is possible to load the
+outputs from this simulation (both linear and nonlinear) into a 
+`GKOutput` object.
+
+.. literalinclude:: example_read_gs2_output.py
+
+More information about the outputs can be found `../howtos/linear_outputs` and `../howtos/read_nonlinear_data`
+
