### 2023-11-23 c1f98d4
Update examples and add contributing guide
