### 2023-11-23 3f52f0c
More verbose README for reference values
