### 2023-11-23 0178b92
Clarify definitions a little
