+++ docs/examples/read_output.rst
@@ -9,5 +9,5 @@ outputs from this simulation (both linear and nonlinear) into a
 
 .. literalinclude:: example_read_gs2_output.py
 
-More information about the outputs can be found `../howtos/linear_outputs` and `../howtos/read_nonlinear_data`
+More information about the outputs can be found `sec-linear-outputs` and `sec-read-nonlinear-data`
 
