### 2023-11-23 #5866 @bpatel2107
@the-rccg I've added some actual examples to the docs [here ](https://pyrokinetics.readthedocs.io/en/latest/examples/index.html) and made the other changes you've suggested. I've also added community guidelines [here](https://pyrokinetics.readthedocs.io/en/latest/developer_guide/contributing.html) and updated the README to be clearer on the definitions.
