### 2023-11-24 1d79e16
Remove larger TRANSP netCDF to save space
