### 2023-11-24 1fdb1da
Remove larger TRANSP netCDF to save space
