### 2023-11-24 bpatel2107/pyrokinetics#262 @bpatel2107
I'd like to jump back to this. We've addressed all the major issues apart from the final issue from the first reviewer about the Docker file. I am leaning more towards it not being necessary but if it is easy to do then I'm not opposed to it. I'm not sure we have any more installation issues (barring the pygacode stuff)

To be honest I'm not really sure what a docker file does and from what I do know how it would benefit  pyro.
