### 2023-11-24 #5866 @bpatel2107
@rogeriojorge
- We use template files from TRANSP in our testing but we did have two which was not really needed so we have removed one of the TRANSP files removing 46MB of required storage
-  `pip install -e .[tests]` may have failed due to an older version of pip. The `README` now suggests to update pip to the latest version
- Those particular tests have been adjusted as they were failing in our CI (we think a change in a scipy routine was the cause but this is now fixed)
- The cause of the error was due to pyrokinetics not being able to automatically identify the file as a GS2 input file. A clearer error message is now provided

```
Traceback (most recent call last):
  File "/home/ir-pate3/test_pyro_install/pyrokinetics/src/pyrokinetics/file_utils.py", line 136, in _chain_exceptions
    _chain_exceptions(*excs)
  File "/home/ir-pate3/test_pyro_install/pyrokinetics/src/pyrokinetics/file_utils.py", line 140, in _chain_exceptions
    raise exc
  File "/home/ir-pate3/test_pyro_install/pyrokinetics/src/pyrokinetics/file_utils.py", line 160, in type
    return super().type(key)
  File "/home/ir-pate3/test_pyro_install/pyrokinetics/src/pyrokinetics/factory.py", line 39, in type
    raise KeyError(
KeyError: "'gs2Input-linear.in' is not recognised as a type of FileReader"

Traceback (most recent call last):
  File "/home/ir-pate3/test_pyro_install/pyrokinetics/src/pyrokinetics/file_utils.py", line 136, in _chain_exceptions
    _chain_exceptions(*excs)
  File "/home/ir-pate3/test_pyro_install/pyrokinetics/src/pyrokinetics/file_utils.py", line 138, in _chain_exceptions
    raise exc from chain
  File "/home/ir-pate3/test_pyro_install/pyrokinetics/src/pyrokinetics/file_utils.py", line 163, in type
    return super().type(self._infer_file_type(key))
  File "/home/ir-pate3/test_pyro_install/pyrokinetics/src/pyrokinetics/file_utils.py", line 191, in _infer_file_type
    raise FileInferenceException(excs)
pyrokinetics.file_utils.FileInferenceException: Could not infer file type. The following exceptions were raised:
+ 'CGYRO' raised --> RuntimeError("Couldn't read CGYRO file. Is the format correct?")
+ 'GENE' raised --> ValueError(" Unable to verify joss.in as a GENE file. The following keys are required: 'general', 'geometry', 'box' ")
+ 'GS2' raised --> ValueError(" Unable to verify joss.in as a GS2 file. The following keys are required: 'knobs', 'parameters', 'theta_grid_knobs', 'theta_grid_eik_knobs', 'theta_grid_parameters', 'species_knobs', 'kt_grids_knobs' ")
+ 'TGLF' raised --> ValueError(" Unable to verify joss.in as a TGLF file. The following keys are required: 'rmin_loc', 'rmaj_loc', 'nky' ")
+ '_test' raised --> RuntimeError('First line of file should say Hello world!.')

Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/ir-pate3/test_pyro_install/pyrokinetics/src/pyrokinetics/pyro.py", line 176, in __init__
    self.read_gk_file(gk_file, gk_code, norms=self.norms)
  File "/home/ir-pate3/test_pyro_install/pyrokinetics/src/pyrokinetics/pyro.py", line 856, in read_gk_file
    gk_input = read_gk_input(gk_file, file_type=gk_code)
  File "/home/ir-pate3/test_pyro_install/pyrokinetics/src/pyrokinetics/gk_code/gk_input.py", line 235, in read_gk_input
    gk_input = GKInput._factory(file_type if file_type is not None else path)
  File "/home/ir-pate3/test_pyro_install/pyrokinetics/src/pyrokinetics/factory.py", line 66, in __call__
    return self.create(key, *args, **kwargs)
  File "/home/ir-pate3/test_pyro_install/pyrokinetics/src/pyrokinetics/factory.py", line 45, in create
    return self.type(key)(*args, **kwargs)
  File "/home/ir-pate3/test_pyro_install/pyrokinetics/src/pyrokinetics/file_utils.py", line 174, in type
    _chain_exceptions(RuntimeError(msg), infer_error, key_error)
  File "/home/ir-pate3/test_pyro_install/pyrokinetics/src/pyrokinetics/file_utils.py", line 138, in _chain_exceptions
    raise exc from chain
RuntimeError: Unable to determine the type of 'GKInput' from the input 'gs2Input-linear.in'. If you supplied a file name, please check the exceptions raised above from each GKInput reader, and try fixing any reported issues. It may also help to specify the file type explicitly.
```

Here the missing namelist are shown. When actually specifying the type of code as follows `pyro = Pyro(gk_file="gs2Input-linear.in", gk_code="GS2")` we come to the root issue which is that pyrokinetics does not support numerical equilibria yet, only analytic ones (Miller, MXH, MillerTurnbull etc.). This of course can be added in the future.

```
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/ir-pate3/test_pyro_install/pyrokinetics/src/pyrokinetics/pyro.py", line 176, in __init__
    self.read_gk_file(gk_file, gk_code, norms=self.norms)
  File "/home/ir-pate3/test_pyro_install/pyrokinetics/src/pyrokinetics/pyro.py", line 871, in read_gk_file
    self.local_geometry = self.gk_input.get_local_geometry()
  File "/home/ir-pate3/test_pyro_install/pyrokinetics/src/pyrokinetics/gk_code/gs2.py", line 148, in get_local_geometry
    raise NotImplementedError(
NotImplementedError: GS2 equilibrium option file not implemented
```

- We are not sure a Docker version of pyrokinetics is quite necessary just yet. We have improved the documentation with some examples and made the installation via pip as simple as possible. Do you have anything you'd specifically want from a Docker (I'm personally not too familiar with them)
