+++ Dockerfile
@@ -1,7 +1,8 @@
 from python:3.9
 
+RUN apt-get update
+RUN apt-get -y install gfortran
 
 COPY ./ /pyrokinetics
 RUN cd /pyrokinetics && pip install .
 
-ENTRYPOINT python3
