### 2023-11-28 107c70a
Run ipython by default
