### 2023-11-29 c737673
Update gk_output.py to fix issue #287
