### 2023-11-29 412ed8a
Update gs2.py to fix calculation of eigenvalues in non-wstar_units runs
