### 2023-11-29 7b1b615
Add Docker explanation to README
