### 2023-11-29 ae421b6
Add ideal_ball solver and example
