+++ src/pyrokinetics/diagnostics/__init__.py
@@ -1,6 +1,8 @@
 import numpy as np
 import xrft
 from scipy.interpolate import RectBivariateSpline
+from scipy.sparse.linalg import eigs
+from scipy.integrate import simps
 
 from ..pyro import Pyro
 
@@ -21,11 +23,6 @@ class Diagnostics:
     """
 
     def __init__(self, pyro: Pyro):
-        if pyro.gk_output is None:
-            raise RuntimeError(
-                "Diagnostics: Please load gk output files (Pyro.load_gk_output)"
-                " before using any diagnostic"
-            )
         self.pyro = pyro
 
     def poincare(
@@ -75,6 +72,12 @@ class Diagnostics:
         RuntimeError: in case of linear simulation
         """
         import pint_xarray  # noqa
+
+        if self.pyro.gk_output is None:
+            raise RuntimeError(
+                "Diagnostics: Please load gk output files (Pyro.load_gk_output)"
+                " before using any diagnostic"
+            )
 
         if self.pyro.gk_code not in ["CGYRO", "GS2", "GENE"]:
             raise NotImplementedError(
@@ -248,3 +251,232 @@ class Diagnostics:
             )
         )
         return np.real(value)
+
+    def gs2_geometry_terms(self, ntheta_multiplier: int = 10):
+        nperiod = self.pyro.numerics.nperiod
+        ntheta = self.pyro.numerics.ntheta * ntheta_multiplier
+
+        theta_max = ((2 * nperiod) - 1) * np.pi
+        theta_even = np.linspace(-theta_max, theta_max, ntheta)
+
+        self.pyro.load_metric_terms(theta=theta_even)
+
+        metric = self.pyro.metric_terms
+
+        theta = metric.regulartheta
+
+        dpdrho = metric.mu0dPdr
+        dpsidrho = metric.dpsidr
+
+        # Metric tensor terms
+        g_ra_contr = metric.field_aligned_contravariant_metric("r", "alpha")
+        g_ra = metric.field_aligned_contravariant_metric("r", "alpha")
+        g_aa = metric.field_aligned_contravariant_metric("alpha", "alpha")
+
+        g_at = metric.field_aligned_covariant_metric("alpha", "theta")
+
+        grad_alpha = np.sqrt(g_aa)
+
+        g_tt = metric.field_aligned_covariant_metric("theta", "theta")
+        dg_tt_dr = metric.dg_theta_theta_dr
+        dg_tt_dtheta = metric.dg_theta_theta_dtheta
+
+        g_rt = metric.field_aligned_covariant_metric("r", "theta")
+        g_rr = metric.field_aligned_contravariant_metric("r", "r")
+        g_rr_tor = metric.toroidal_contravariant_metric("r", "r")
+
+        dg_zz_drho = metric.dg_zeta_zeta_dr
+        dg_zz_dtheta = metric.dg_zeta_zeta_dtheta
+
+        g_tt_tor = metric.toroidal_covariant_metric("theta", "theta")
+        g_zz_tor = metric.toroidal_covariant_metric("zeta", "zeta")
+
+        # Jacobian and derivatives
+        jacob = metric.Jacobian
+        djacob_drho = metric.dJacobian_dr
+        djacob_dtheta = metric.dJacobian_dtheta
+
+        # R and derivatives
[…]
