### 2023-11-29 433c5e6
Tests for gs2 diagnostics
