### 2023-11-30 #5866 @bpatel2107
We decided that a docker was straightforward to add so it is now included and there's an update to the README with instructions on how to use it.
