### 2023-12-01 96ad632
Use tqdm to speed up ideal ballooning calculation
