+++ pyproject.toml
@@ -46,6 +46,7 @@ dependencies = [
     "gitpython >= 3.1",
     "jetto-tools",
     "pygacode",
+    "tqdm",
 ]
 
 [project.optional-dependencies]
