### 2023-12-01 d2266b3
Use tqdm to speed up ideal ballooning calculation
