### 2023-12-01 bec216e
Set OMP_NUM_THREADS in example_ideal_ballooning
