### 2023-12-01 7008cd4
Flake8 ignore example_ideal_ballooning
Suppresses warning about imports not being at top of file. This
is needed to set the number of OpenMP threads before libraries
like NumPy and SciPy are imported.
