### 2023-12-04 05d79d6
update docstring and reorganise multiplication
