### 2023-12-05 de048bd
Add theta0 dependance
