### 2023-12-05 55b1d68
Move theta0 into ideal ball method
