### 2023-12-06 a4b7d34
Fix to Dockerfile missing .git dir
Also adds testing dependencies and uses WORKDIR to simplify commands
