### 2023-12-06 eb99724
Edit to COPY command in Dockerfile
