### 2023-12-06 b8f18f3
Remove git calls from Dockerfile
