### 2023-12-06 d9c0f56
Use SETUPTOOLS_SCM_PRETEND_VERSION in Dockerfile
