### 2023-12-06 320334a
Use SETUPTOOLS_SCM_PRETEND_VERSION in Dockerfile
