### 2023-12-06 a21ad6e
Build docker image on push
