### 2023-12-06 f0a9321
initial synthetic highk/dbs diagnostics
