+++ src/pyrokinetics/diagnostics/__init__.py
@@ -6,6 +6,7 @@ from scipy.sparse.linalg import eigs
 
 from ..pyro import Pyro
 
+from .synthetic_highk_dbs import Synthetic_highk_dbs
 
 class Diagnostics:
     """
