### 2023-12-06 e4f4af0
[skip ci] Apply black changes
