### 2023-12-18 #5866 @the-rccg
@kellyrowland I've been able to verify the functionality using version 0.5.0 (as opposed to the current 0.6.1) due to issues with getting pygacode 1.0 running on Windows.

It's all working well with that, great work on the code base! I'm sure this will find good use in the community.

One thing I was not able to confirm is the Diagnostic usage from the Ideal Ballooning Solver (https://pyrokinetics.readthedocs.io/en/latest/examples/ideal_ballooning.html), which might be due to the older version and the example was added after the release of the 0.5.0 version verified, so ignoring it here.

```python
diag = Diagnostics(pyro)  # RuntimeError: Diagnostics: Please load gk output files (Pyro.load_gk_output) before using any diagnostic
```
