### 2023-12-18 #5866 @kellyrowland
thanks very much!

@bpatel2107 could you comment on the above? it looks like we have the checkboxes all filled out, but I want to make sure that the `RuntimeError` has been addressed in the most recent release.
