### 2023-12-19 #5866 @bpatel2107
@kellyrowland ,@the-rccg the ideal ballooning solver was added in v.0.6 so it would not work on6 v0.5. The only diagnostic in v0.5 required the output hence the warning. This warning has been removed in v0.6.

Note we are working with the pygacode team to get their code to work without needing a Fortran compiler so a future release should be fine on Windows. You can run the docker container in the meantime.
