### 2024-01-09 #5866 @bpatel2107
I'm currently on leave until the end of February. The latest release with all these changes is v0.6.1

I'll see if I can upload to zenodo in the meantime (without a laptop so doing everything from my phone)
