### 2024-03-04 b9772b6
Replace "parameters" with "knobs" in GS2
The parameters namelist in GS2 has been deprecated for a while and
v8.2 will be removing this. Currently we attempt to get/set beta and
zeff using parameters. This commit changes that.
