+++ src/pyrokinetics/gk_code/gs2.py
@@ -106,7 +106,6 @@ class GKInputGS2(GKInput, FileReader, file_type="GS2", reads=GKInput):
         # but they are needed by Pyrokinetics
         expected_keys = [
             "knobs",
-            "parameters",
             "theta_grid_knobs",
             "theta_grid_eik_knobs",
             "theta_grid_parameters",
@@ -202,7 +201,7 @@ class GKInputGS2(GKInput, FileReader, file_type="GS2", reads=GKInput):
 
         ne_norm, Te_norm = self.get_ne_te_normalisation()
         beta = (
-            self.data["parameters"]["beta"]
+            self._get_beta()
             * (miller_data["Rmaj"] / r_geo) ** 2
             * ne_norm
             * Te_norm
@@ -274,7 +273,7 @@ class GKInputGS2(GKInput, FileReader, file_type="GS2", reads=GKInput):
 
         if "zeff" in self.data["knobs"]:
             local_species.zeff = self.data["knobs"]["zeff"] * ureg.elementary_charge
-        elif "zeff" in self.data["parameters"]:
+        elif "parameters" in self.data.keys() and "zeff" in self.data["parameters"]:
             local_species.zeff = (
                 self.data["parameters"]["zeff"] * ureg.elementary_charge
             )
@@ -424,7 +423,7 @@ class GKInputGS2(GKInput, FileReader, file_type="GS2", reads=GKInput):
         r_geo = self.data["theta_grid_parameters"].get("r_geo", Rmaj)
 
         ne_norm, Te_norm = self.get_ne_te_normalisation()
-        beta = self.data["parameters"]["beta"] * (Rmaj / r_geo) ** 2 * ne_norm * Te_norm
+        beta = self._get_beta() * (Rmaj / r_geo) ** 2 * ne_norm * Te_norm
         numerics_data["beta"] = beta * ureg.beta_ref_ee_B0
 
         numerics_data["gamma_exb"] = (
@@ -542,7 +541,7 @@ class GKInputGS2(GKInput, FileReader, file_type="GS2", reads=GKInput):
         self.data["knobs"]["zeff"] = local_species.zeff
 
         beta_ref = local_norm.gs2.beta if local_norm else 0.0
-        self.data["parameters"]["beta"] = (
+        self.data["knobs"]["beta"] = (
             numerics.beta if numerics.beta is not None else beta_ref
         )
 
@@ -757,6 +756,17 @@ class GKOutputReaderGS2(FileReader, file_type="GS2", reads=GKOutput):
             pass
         else:
             raise RuntimeError(f"file '{filename}' missing expected GS2 attributes")
+
+    def _get_beta(self):
+        """
+        Small helper to wrap up logic required to get beta from the input
+        consistent with logic across versions of GS2.
+        """
+        has_parameters = "parameters" in self.data.keys()
+        beta_default = 0.0
+        if has_parameters:
+            beta_default = self.data["parameters"].get("beta", 0.0)
+        return self.data["knobs"].get("beta", beta_default)
 
     @staticmethod
     def infer_path_from_input_file(filename: PathLike) -> Path:
