### 2024-03-04 #5866 @kellyrowland
@bpatel2107 it looks like there are several discrepancies between the package and the archive that will need to be aligned:

- there is a typo in the DOI title, "Pyrokinetivs" should be "Pyrokinetics"
- the author list should be the same on the manuscript and the archive; there's only one author listed on the archive
- the license listed on the archive doesn't match the license in the code repo

I think you should be able to make these changes by editing the metadata of the record on Zenodo (see https://help.zenodo.org/faq/#ctechnical-1).
