### 2024-03-05 add4df5
Update MG affiliation
